#2049 Moved two functsion from lfric_extract_driver_creator_test... into base_driver_creator_test
